+++ suubi-solutions/src/App.js
@@ -1,12 +1,12 @@
 
 import './App.css';
-import Navbar from './Landing/landing';
-import Footer from './Footer/footer';
+
+import Footer from './Footer/index';
 
 function App() {
   return (
     <div className="App">
-      <Navbar/>
+      
       <Footer/>
     </div>
   );
