--- suubi-solutions/src/Footer/footer.css
@@ -1,115 +1,0 @@
-.link a{
-text-decoration: none;
-color: white;
-}
-footer{
-display: flex;
-font-size: 16px;
-gap: 380px;
-background-color: #143A62;
-color: white;
-padding-left: 10px;
-}
-.Icons{
-width: 29.64px;
-height: 24.08px;
-margin: 3px;
-
-}
-.logos{
-width: 90px;
-height: 90px;
-margin-left: 50px;
-margin-top: 50px;
-        
-}
-@media screen and (min-width:280px) and (max-width:480px){
-    footer{
-        
-    flex-direction: column;
-    font-size: 16px;
-    gap: 3px;
-    background-color: #143A62;
-    color: white;
-    padding-left: 10px;
-    }
-    
-    .logos{
-      
-        margin-left: 130px;
-        
-            
-    }
-}
-@media screen and (min-width:481px) and (max-width:681px){
-    footer{
-        
-        flex-direction: column;
-        font-size: 16px;
-        gap: 3px;
-        background-color: #143A62;
-        color: white;
-        padding-left: 10px;
-        }
-        
-        .logos{
-          
-            margin-left: 130px;
-            
-                
-        }
-}
-@media screen  and (min-width:682px) and (max-width:882px){
-    footer{
-        
-        flex-direction: column;
-        font-size: 16px;
-        gap: 3px;
-        background-color: #143A62;
-        color: white;
-        padding-left: 10px;
-        }
-        
-        .logos{
-          
-            margin-left: 130px;
-            
-                
-        }
-}
-@media screen  and (min-width:883px) and (max-width:1083px){
-    footer{
-        
-        flex-direction: column;
-        font-size: 16px;
-        gap: 3px;
-        background-color: #143A62;
-        color: white;
-        padding-left: 10px;
-        }
-        
-        .logos{
-          
-            margin-left: 130px;
-            
-                
-        }
-}
-@media screen  and (min-width:1084px) and (max-width:1284px){
-    footer{
-        
-        flex-direction: column;
-        font-size: 16px;
-        gap: 3px;
-        background-color: #143A62;
-        color: white;
-        padding-left: 10px;
-        }
-        
-        .logos{
-          
-            margin-left: 130px;
-            
-                
-        }
-}
